Move ICMP packet creation and parsing out of Pinger implementation
